Preventing a data race.
Resize() can get called from a goroutine that is not the primary client
of the termbox, e.g. an ssh session recieving a window change event.
This change places a mutex around the newW/newH variables so that there
won't be a race between witing them on the ssh session goroutine and
reading them in the termbox client.  This also changes PollEvent and
PollRawEvent to not change the backbuffer/frontbuffer sizes when a
window size change event is recieved, as some termbox clients process
events in a seperate goroutine from the gorotuine that is writing to
the the buffers and flushing the buffers.